requirements.txt
matplotlib==3.6.2
numpy==1.23.4
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
requests==2.28.1
scikit_learn==1.1.3
streamlit==1.14.1
tensorflow==2.11.0
tensorflow_gpu==2.10.0
torch==1.9.0+cu102
torchvision==0.10.0+cu102
